+++ Source/Scene/Cesium3DTileStyle.js
@@ -250,13 +250,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     show : '(regExp("^Chest").test(${County})) && (${YearBuilt} >= 1970)'
          * });
-         * style.show.evaluate(frameState, feature); // returns true or false depending on the feature's properties
+         * style.show.evaluate(feature); // returns true or false depending on the feature's properties
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override show expression with a custom function
          * style.show = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return true;
          *     }
          * };
@@ -319,13 +319,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     color : '(${Temperature} > 90) ? color("red") : color("white")'
          * });
-         * style.color.evaluateColor(frameState, feature, result); // returns a Cesium.Color object
+         * style.color.evaluateColor(feature, result); // returns a Cesium.Color object
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override color expression with a custom function
          * style.color = {
-         *     evaluateColor : function(frameState, feature, result) {
+         *     evaluateColor : function(feature, result) {
          *         return Cesium.Color.clone(Cesium.Color.WHITE, result);
          *     }
          * };
@@ -381,13 +381,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     pointSize : '(${Temperature} > 90) ? 2.0 : 1.0'
          * });
-         * style.pointSize.evaluate(frameState, feature); // returns a Number
+         * style.pointSize.evaluate(feature); // returns a Number
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override pointSize expression with a custom function
          * style.pointSize = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return 1.0;
          *     }
          * };
@@ -690,13 +690,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     font : '(${Temperature} > 90) ? "30px Helvetica" : "24px Helvetica"'
          * });
-         * style.font.evaluate(frameState, feature); // returns a String
+         * style.font.evaluate(feature); // returns a String
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override font expression with a custom function
          * style.font = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return '24px Helvetica';
          *     }
          * };
@@ -738,13 +738,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     labelStyle : '(${Temperature} > 90) ? ' + LabelStyle.FILL_AND_OUTLINE + ' : ' + LabelStyle.FILL
          * });
-         * style.labelStyle.evaluate(frameState, feature); // returns a LabelStyle
+         * style.labelStyle.evaluate(feature); // returns a LabelStyle
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override labelStyle expression with a custom function
          * style.labelStyle = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return LabelStyle.FILL;
          *     }
          * };
@@ -786,13 +786,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     labelText : '(${Temperature} > 90) ? ">90" : "<=90"'
          * });
-         * style.labelText.evaluate(frameState, feature); // returns a String
+         * style.labelText.evaluate(feature); // returns a String
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override labelText expression with a custom function
          * style.labelText = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return 'Example label text';
          *     }
          * };
@@ -882,7 +882,7 @@ define([
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override backgroundPadding expression with a string
          * style.backgroundPadding = 'vec2(5.0, 7.0)';
-         * style.backgroundPadding.evaluate(frameState, feature); // returns a Cartesian2
+         * style.backgroundPadding.evaluate(feature); // returns a Cartesian2
          */
         backgroundPadding : {
             get : function() {
@@ -969,7 +969,7 @@ define([
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override scaleByDistance expression with a string
          * style.scaleByDistance = 'vec4(1.5e2, 2.0, 1.5e7, 0.5)';
-         * style.scaleByDistance.evaluate(frameState, feature); // returns a Cartesian4
+         * style.scaleByDistance.evaluate(feature); // returns a Cartesian4
          */
         scaleByDistance : {
             get : function() {
@@ -1008,7 +1008,7 @@ define([
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override translucencyByDistance expression with a string
          * style.translucencyByDistance = 'vec4(1.5e2, 1.0, 1.5e7, 0.2)';
-         * style.translucencyByDistance.evaluate(frameState, feature); // returns a Cartesian4
+         * style.translucencyByDistance.evaluate(feature); // returns a Cartesian4
          */
         translucencyByDistance : {
             get : function() {
@@ -1047,7 +1047,7 @@ define([
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override distanceDisplayCondition expression with a string
          * style.distanceDisplayCondition = 'vec2(0.0, 5.5e6)';
-         * style.distanceDisplayCondition.evaluate(frameState, feature); // returns a Cartesian2
+         * style.distanceDisplayCondition.evaluate(feature); // returns a Cartesian2
          */
         distanceDisplayCondition : {
             get : function() {
@@ -1230,13 +1230,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     image : '(${Temperature} > 90) ? "/url/to/image1" : "/url/to/image2"'
          * });
-         * style.image.evaluate(frameState, feature); // returns a String
+         * style.image.evaluate(feature); // returns a String
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override image expression with a custom function
          * style.image = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return '/url/to/image';
          *     }
          * };
@@ -1278,7 +1278,7 @@ define([
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override disableDepthTestDistance expression with a string
          * style.disableDepthTestDistance = '1000.0';
-         * style.disableDepthTestDistance.evaluate(frameState, feature); // returns a Number
+         * style.disableDepthTestDistance.evaluate(feature); // returns a Number
          */
         disableDepthTestDistance : {
             get : function() {
@@ -1317,13 +1317,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     horizontalOrigin : HorizontalOrigin.LEFT
          * });
-         * style.horizontalOrigin.evaluate(frameState, feature); // returns a HorizontalOrigin
+         * style.horizontalOrigin.evaluate(feature); // returns a HorizontalOrigin
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override horizontalOrigin expression with a custom function
          * style.horizontalOrigin = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return HorizontalOrigin.CENTER;
          *     }
          * };
@@ -1365,13 +1365,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     verticalOrigin : VerticalOrigin.TOP
          * });
-         * style.verticalOrigin.evaluate(frameState, feature); // returns a VerticalOrigin
+         * style.verticalOrigin.evaluate(feature); // returns a VerticalOrigin
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override verticalOrigin expression with a custom function
          * style.verticalOrigin = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return VerticalOrigin.CENTER;
          *     }
          * };
@@ -1413,13 +1413,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     labelHorizontalOrigin : HorizontalOrigin.LEFT
          * });
-         * style.labelHorizontalOrigin.evaluate(frameState, feature); // returns a HorizontalOrigin
+         * style.labelHorizontalOrigin.evaluate(feature); // returns a HorizontalOrigin
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override labelHorizontalOrigin expression with a custom function
          * style.labelHorizontalOrigin = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return HorizontalOrigin.CENTER;
          *     }
          * };
@@ -1461,13 +1461,13 @@ define([
          * var style = new Cesium3DTileStyle({
          *     labelVerticalOrigin : VerticalOrigin.TOP
          * });
-         * style.labelVerticalOrigin.evaluate(frameState, feature); // returns a VerticalOrigin
+         * style.labelVerticalOrigin.evaluate(feature); // returns a VerticalOrigin
          *
          * @example
          * var style = new Cesium.Cesium3DTileStyle();
          * // Override labelVerticalOrigin expression with a custom function
          * style.labelVerticalOrigin = {
-         *     evaluate : function(frameState, feature) {
+         *     evaluate : function(feature) {
          *         return VerticalOrigin.CENTER;
          *     }
          * };
@@ -1503,7 +1503,7 @@ define([
          *         description : '"Building id ${id} has height ${Height}."'
          *     }
          * });
-         * style.meta.description.evaluate(frameState, feature); // returns a String with the substituted variables
+         * style.meta.description.evaluate(feature); // returns a String with the substituted variables
          */
         meta : {
             get : function() {